Split grpc/util.go into client and server.
Having files or packages named util is not great, because they wind up
attracting lots of small, unrelated functionality.
